+++ lib/src/domain/respositories/package_repository.dart
@@ -2,5 +2,4 @@ import 'package:cura/src/domain/models/package_info.dart';
 
 abstract class PackageRepository {
   Future<PackageInfo> getPackageInfo(String packageName);
-  Future<List<PackageInfo>> getPackagesFromPubspec(String pubspecPath);
 }